gestion d'erreur ok

+++ app/controller/jobController.js
@@ -3,77 +3,79 @@ const jobDatamapper = require('../datamapper/jobDatamapper');
 const jobController ={
 
 
-    async addJobProject (req, res, next) {
+    async addJobProject (req, res) {
         try {
             const projectId= req.params.id;
+            if (!projectId){
+                throw new MainError('missing parameter', req, res, 400);
+            };
             const jobName = req.body.job
             const jobfinded = await jobDatamapper.getJobId(jobName);
-            const result = await jobDatamapper.addJob (projectId ,jobfinded.id);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!jobfinded){
+                throw new MainError('This job does not exists', req, res, 400);
             };
-
+            const result = await jobDatamapper.addJob (projectId ,jobfinded.id);
+            if (!result){
+                throw new MainError('This job was not adding', req, res, 404);
+            };
             return res.status(204).json(result);
-
         } catch (error) {
-
-        next(error);
-
+        console.error(error);
         };
     },
 
-    async deleteJobProject (req, res, next) {
+    async deleteJobProject (req, res) {
         try {
             const tableId = req.body.id_project_has_job;
-            const result = await jobDatamapper.deleteJobProject (tableId);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!tableId){
+                throw new MainError('missing parameter', req, res, 400);
             };
-
+            const result = await jobDatamapper.deleteJobProject (tableId);
+            if (!result){
+                throw new MainError('This job was not delete', req, res, 404);
+            };
             return res.status(204).json(result);
-
         } catch (error) {
-
-        next(error);
-
+        console.error(error);
         };
     },
         
 
-    async addJobUser (req, res, next) {
+    async addJobUser (req, res) {
         try {
             const userId= req.params.id;
+            if (!userId){
+                throw new MainError('missing parameter', req, res, 400);
+            };
             const jobName = req.body.job
             console.log('ici 2');
             const jobfinded = await jobDatamapper.getJobId(jobName);
-            const result = await jobDatamapper.addJobUser (userId ,jobfinded.id);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!jobfinded){
+                throw new MainError('This job does not exists', req, res, 400);
             };
-
+            const result = await jobDatamapper.addJobUser (userId ,jobfinded.id);
+            if (!result){
+                throw new MainError('This job was not adding', req, res, 404);
+            };
             return res.status(204).json(result);;
-
         } catch (error) {
-
-        next(error);
-
+        console.error(error);
         };
     },
 
-    async deleteJobUser (req, res, next) {
+    async deleteJobUser (req, res) {
         try {
             const userId= req.params.id;
-            const result = await jobDatamapper.deleteJobUser(userId);
-            if (result === null || result === undefined){
-                return res.status(404).json({ message: "This job does not exists !"});
+            if (!userId){
+                throw new MainError('missing parameter', req, res, 400);
             };
-
+            const result = await jobDatamapper.deleteJobUser(userId);
+            if (!result){
+                throw new MainError('This job was not delete', req, res, 404);
+            };
             return res.status(204).json(result);
-
         } catch (error) {
-
-            next(error);
-
+            console.error(error);
         };
     }
